requirements.txt
asttokens==2.4.1
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
MLB-StatsAPI==1.7.2
nest-asyncio==1.6.0
packaging==24.0
parso==0.8.4
platformdirs==4.2.2
prompt_toolkit==3.0.45
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-mlb-statsapi==0.5.17
pywin32==306
pyzmq==26.0.3
requests==2.32.3
requests-mock==1.12.1
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
urllib3==2.2.1
wcwidth==0.2.13
